Better displayed errors and new standards!
New standard:
Forms are to have the names of: <letter of the alphabet in ascending
order>_<field name>.
Ex: a_username, b_password.
In order for errors to be sorted because the are sorted in alphabetical
order.

New available constant: noErrorMessage which is to be used to create
errors with only keys (display field feedback) but no messages.
Example use: c.Validation.Error(noErrorMessage).Key(“b_password”)
Internally, this will be verified with the added template function of
“ismessage”.
Usage:
{{if ismessage $e}}
    // Display output
{{end}}

Gonna sleep happy tonight ;)